Improve Airtable connector DX/UX based on review feedback
- Extract manifest to separate manifest.go (matches Slack/Salesforce pattern)
- Improve parameter descriptions with examples and where-to-find hints
- Expand error mapping: TABLE_NOT_FOUND, VIEW_NOT_FOUND, FIELD_NOT_FOUND,
  INVALID_FILTER_BY_FORMULA, and more with actionable messages
- Improve auth error messages (clear guidance vs raw JSON dump)
- Add parseAndValidate generic helper to reduce boilerplate across actions
- Fix URL encoding in list_bases offset parameter
- Collapse duplicate deleteRecordSummary/deleteRecordEntry types
- Add comprehensive README.md with actions, error mapping, and file structure
- Clarify search_records description as convenience wrapper
